+++ src/compile/mod.rs
@@ -798,6 +798,13 @@ code:
                     url = format!("/{url}");
                 }
                 url = format!("github.com{url}");
+                self.emit_diagnostic(
+                    "Implicit GitHub URLs are deprecated and \
+                    will be removed in the future. Prefix the \
+                    URL with `github.com/` or `https://github.com/`.",
+                    DiagnosticKind::Warning,
+                    span.clone(),
+                );
             }
             if !(url.starts_with("https://") || url.starts_with("http://")) {
                 url = format!("https://{url}");