+++ src/algorithm/zip.rs
@@ -63,6 +63,7 @@ fn prim_mon_fast_fn(prim: Primitive, span: usize) -> Option<ValueMonFn> {
             Ok(v)
         }),
         Box => spanned_mon_fn(span, |v, d, _| Ok(v.box_depth(d))),
+        First => spanned_mon_fn(span, |v, d, env| v.first_depth(d, env)),
         _ => return None,
     })
 }
@@ -91,6 +92,7 @@ fn impl_prim_mon_fast_fn(prim: ImplPrimitive, span: usize) -> Option<ValueMonFn>
             v.sort_down_depth(d);
             Ok(v)
         }),
+        Last => spanned_mon_fn(span, |v, d, env| v.last_depth(d, env)),
         _ => return None,
     })
 }